adding jwt functionality to user svc
