+++ user-service/src/main/java/com/jrsmiffy/jara3/userservice/model/Role.java
@@ -1,0 +1,8 @@
+package com.jrsmiffy.jara3.userservice.model;
+
+public enum Role {
+    USER,
+    ADMIN
+}
+
+// TODO: extract this to a common-utils service
