+++ user-service/src/main/java/com/jrsmiffy/jara3/userservice/model/UserResponse.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @AllArgsConstructor
 public class UserResponse {
 
-    private Optional<User> user;
+    private Optional<AppUser> user;
     private String response;
 }
